1. Change table names and column names 2. Handle principle and score tags in Assessment level instead of Project level 3. Remove unnecessary functions in Project model
