+++ app/Models/PrincipleAssessment.php
@@ -6,7 +6,7 @@ use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\Pivot;
 
-class PrincipleProject extends Model
+class PrincipleAssessment extends Model
 {
 
     public $incrementing = true;
